+++ add_changelog.sh
@@ -19,16 +19,20 @@ DE_CHANGES=$(cat << EOF
 - Speicher- und Ladebuttons sind nun getrennt.
 - Der „Speichern“-Button im Spielbildschirm wird immer wieder aktiviert.
 - Ein Lade-Indikator wird nun im Spielfeld angezeigt, wenn die KI rechnet.
+- Option in den Einstellungen, ob jedes mal eine neue Karte generiert wird
+- Sound Effekte für Roboterbewegungen
 EOF
 )
 
 # Englische Übersetzung
 EN_CHANGES=$(cat << EOF
-- add cancle button if AI takes too long for solving after 10s
+- add cancel button if AI takes too long for solving after 10s
 - Game history is now automatically saved before opening the save screen.
 - Save and load buttons are now separate.
 - The "Save" button in the game screen is always reactivated.
 - A loading indicator is now displayed in the game field when the AI is calculating.
+- Settings option to choose if a new map is generated each game
+- Added sound effects for robot movement
 EOF
 )